+++ share_handler/lib/src/share_handler.dart
@@ -1,5 +1,6 @@
 import 'package:share_handler_platform_interface/share_handler_platform_interface.dart';
 
 class ShareHandler {
+  /// Gets ShareHandlerPlatform instance to use.
   static ShareHandlerPlatform get instance => ShareHandlerPlatform.instance;
 }