+++ share_handler/CHANGELOG.md
@@ -1,2 +1,4 @@
+# 0.0.3
+Fix for error getting initial media
 # 0.0.2
 Initial release of this plugin.