+++ share_handler_platform_interface/CHANGELOG.md
@@ -1,3 +1,5 @@
+# 0.0.3
+Fix for error getting initial media
 # 0.0.2
 Updated License
 Code cleanup